havacilikda_yapayzeka/birinci_gorev/requirements.txt
# ============================================================
# Havacılıkta Yapay Zekâ Yarışması - Python Bağımlılıkları
# ============================================================

# --- Derin Öğrenme ---
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0        # YOLOv8+

# --- Görüntü İşleme ---
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0     # Veri augmentasyonu

# --- Bilimsel Hesaplama ---
numpy>=1.24.0
scipy>=1.11.0

# --- Veri İşleme ---
pandas>=2.0.0
PyYAML>=6.0.0

# --- Görselleştirme ---
matplotlib>=3.7.0
seaborn>=0.12.0

# --- Utility ---
tqdm>=4.65.0              # Progress bar
loguru>=0.7.0             # Logging
click>=8.1.0              # CLI arayüzü

# --- Geliştirme Araçları ---
jupyter>=1.0.0
ipykernel>=6.0.0
black>=23.0.0             # Kod formatlama
pytest>=7.4.0             # Test çerçevesi
